Fix account and dev settings unit tests
