+++ app/src/test/java/com/babylon/wallet/android/presentation/accountsettings/AccountSettingsViewModelTest.kt
@@ -8,6 +8,8 @@ import com.babylon.wallet.android.presentation.StateViewModelTest
 import com.babylon.wallet.android.presentation.account.settings.ARG_ACCOUNT_SETTINGS_ADDRESS
 import com.babylon.wallet.android.presentation.account.settings.AccountSettingsViewModel
 import com.babylon.wallet.android.presentation.account.settings.Event
+import com.babylon.wallet.android.utils.AppEvent
+import com.babylon.wallet.android.utils.AppEventBus
 import io.mockk.Runs
 import io.mockk.coEvery
 import io.mockk.coVerify
@@ -15,7 +17,10 @@ import io.mockk.every
 import io.mockk.just
 import io.mockk.mockk
 import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.first
+import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOf
+import kotlinx.coroutines.test.TestScope
 import kotlinx.coroutines.test.advanceUntilIdle
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
@@ -35,6 +40,8 @@ internal class AccountSettingsViewModelTest : StateViewModelTest<AccountSettings
     private val changeEntityVisibilityUseCase = mockk<ChangeEntityVisibilityUseCase>()
     private val sampleProfile = sampleDataProvider.sampleProfile()
     private val sampleAddress = sampleProfile.currentNetwork!!.accounts.first().address
+    private val eventBus = mockk<AppEventBus>()
+    private val sampleTxId = "txId1"
 
     override fun initVM(): AccountSettingsViewModel {
         return AccountSettingsViewModel(
@@ -42,7 +49,9 @@ internal class AccountSettingsViewModelTest : StateViewModelTest<AccountSettings
             getProfileUseCase,
             renameAccountDisplayNameUseCase,
             savedStateHandle,
-            changeEntityVisibilityUseCase
+            changeEntityVisibilityUseCase,
+            TestScope(),
+            eventBus
         )
     }
 
@@ -50,9 +59,11 @@ internal class AccountSettingsViewModelTest : StateViewModelTest<AccountSettings
     override fun setUp() {
         super.setUp()
         every { getFreeXrdUseCase.getFaucetState(any()) } returns flowOf(FaucetState.Available(true))
+        coEvery { getFreeXrdUseCase(any()) } returns Result.success(sampleTxId)
         every { getProfileUseCase() } returns flowOf(sampleDataProvider.sampleProfile())
         every { savedStateHandle.get<String>(ARG_ACCOUNT_SETTINGS_ADDRESS) } returns sampleAddress
         coEvery { changeEntityVisibilityUseCase.hideAccount(any()) } just Runs
+        coEvery { eventBus.sendEvent(any()) } just Runs
     }
 
     @Test
@@ -85,4 +96,42 @@ internal class AccountSettingsViewModelTest : StateViewModelTest<AccountSettings
         }
     }
 
+    @Test
+    fun `initial state is correct when free xrd enabled`() = runTest {
+        val vm = vm.value
+        advanceUntilIdle()
+        val state = vm.state.first()
+        assert(state.faucetState is FaucetState.Available)
+    }
+
+    @Test
+    fun `initial state is correct when free xrd not enabled`() = runTest {
+        every { getFreeXrdUseCase.getFaucetState(any()) } returns flow { emit(FaucetState.Available(false)) }
+        val vm = vm.value
+        advanceUntilIdle()
+        val state = vm.state.first()
+        assert(state.faucetState is FaucetState.Available && !(state.faucetState as FaucetState.Available).isEnabled)
+    }
+
+    @Test
+    fun `get free xrd success sets proper state`() = runTest {
+        val vm = vm.value
+        advanceUntilIdle()
+        vm.onGetFreeXrdClick()
+        advanceUntilIdle()
+        coVerify(exactly = 1) { eventBus.sendEvent(AppEvent.RefreshResourcesNeeded) }
+        coVerify(exactly = 1) { getFreeXrdUseCase(sampleAddress) }
+    }
+
+    @Test
+    fun `get free xrd failure sets proper state`() = runTest {
+        coEvery { getFreeXrdUseCase(any()) } returns Result.failure(Exception())
+        val vm = vm.value
+        advanceUntilIdle()
+        vm.onGetFreeXrdClick()
+        advanceUntilIdle()
+        val state = vm.state.first()
+        coVerify(exactly = 1) { getFreeXrdUseCase(sampleAddress) }
+        assert(state.error != null)
+    }
 }
